Refactor the SearchFiltersWrapper so the FeaturedFilters and IconFilters are separated in an obvious way. I added a note about some improvements that could be made to this wrapper in the future, which should be done if we ever need to use this wrapper with _only_ featured filters, as opposed to featured and icon filters together.
